# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "PveSphere is a comprehensive web-based management platform for Proxmox VE (PVE) clusters.",
        "title": "PveSphere API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "PveSphere Support",
            "url": "https://github.com/pvesphere/pvesphere",
            "email": "support@pvesphere.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "paths": {
...
        "/api/v1/dashboard/hotspots": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard模块"
                ],
                "summary": "获取压力和风险焦点",
                "parameters": [
                    {
                        "type": "string",
                        "default": "all",
                        "description": "范围: all 或 cluster",
                        "name": "scope",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "集群ID（当 scope 为 cluster 时使用）",
                        "name": "cluster_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 5,
                        "description": "Top N 数量",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pvesphere_api_v1.DashboardHotspotsResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/dashboard/operations": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard模块"
                ],
                "summary": "获取运行中的操作",
                "parameters": [
                    {
                        "type": "string",
                        "default": "all",
                        "description": "范围: all 或 cluster",
                        "name": "scope",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "集群ID（当 scope 为 cluster 时使用）",
                        "name": "cluster_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pvesphere_api_v1.DashboardOperationsResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/dashboard/overview": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard模块"
                ],
                "summary": "获取全局概览",
                "parameters": [
                    {
                        "type": "string",
                        "default": "all",
                        "description": "范围: all 或 cluster",
                        "name": "scope",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "集群ID（当 scope 为 cluster 时使用）",
                        "name": "cluster_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pvesphere_api_v1.DashboardOverviewResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/dashboard/resources": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard模块"
                ],
                "summary": "获取资源使用率",
                "parameters": [
                    {
                        "type": "string",
                        "default": "all",
                        "description": "范围: all 或 cluster",
                        "name": "scope",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "集群ID（当 scope 为 cluster 时使用）",
                        "name": "cluster_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pvesphere_api_v1.DashboardResourcesResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/dashboard/scopes": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard模块"
                ],
                "summary": "获取可选集群列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pvesphere_api_v1.DashboardScopesResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/pve/access/ticket": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "按照 Proxmox 原生 /api2/json/access/ticket 接口封装，使用用户名/密码获取 ticket 和 CSRFPreventionToken。api_url 从集群表中根据 cluster_id 获取。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PVE认证模块"
                ],
                "summary": "获取 Proxmox 高权限票据（/access/ticket）",
                "parameters": [
                    {
                        "description": "params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pvesphere_api_v1.GetAccessTicketRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pvesphere_api_v1.GetAccessTicketResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "pvesphere_api_v1.DashboardHotspotsData": {
            "type": "object",
            "properties": {
                "cluster_id": {
                    "description": "集群ID（当 scope 为 cluster 时）",
                    "type": "integer"
                },
                "node_hotspots": {
                    "description": "节点热点",
                    "allOf": [
                        {
                            "$ref": "#/definitions/pvesphere_api_v1.NodeHotspots"
                        }
                    ]
                },
                "recent_risks": {
                    "description": "最近风险（24h）",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pvesphere_api_v1.RecentRisk"
                    }
                },
                "scope": {
                    "description": "all 或 cluster",
                    "type": "string",
                    "example": "all"
                },
                "storage_hotspots": {
                    "description": "存储热点",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pvesphere_api_v1.StorageHotspot"
                    }
                },
                "vm_hotspots": {
                    "description": "虚拟机热点",
                    "allOf": [
                        {
                            "$ref": "#/definitions/pvesphere_api_v1.VMHotspots"
                        }
                    ]
                }
            }
        },
        "pvesphere_api_v1.DashboardHotspotsResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/pvesphere_api_v1.DashboardHotspotsData"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "pvesphere_api_v1.DashboardOperationsData": {
            "type": "object",
            "properties": {
                "cluster_id": {
                    "description": "集群ID（当 scope 为 cluster 时）",
                    "type": "integer"
                },
                "items": {
                    "description": "操作明细（可选）",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pvesphere_api_v1.OperationItem"
                    }
                },
                "scope": {
                    "description": "all 或 cluster",
                    "type": "string",
                    "example": "all"
                },
                "summary": {
                    "description": "操作摘要（按类型聚合）",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pvesphere_api_v1.OperationSummary"
                    }
                }
            }
        },
        "pvesphere_api_v1.DashboardOperationsResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/pvesphere_api_v1.DashboardOperationsData"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "pvesphere_api_v1.DashboardOverviewData": {
            "type": "object",
            "properties": {
                "cluster_id": {
                    "description": "集群ID（当 scope 为 cluster 时）",
                    "type": "integer"
                },
                "health": {
                    "description": "健康状态",
                    "allOf": [
                        {
                            "$ref": "#/definitions/pvesphere_api_v1.DashboardOverviewHealth"
                        }
                    ]
                },
                "scope": {
                    "description": "all 或 cluster",
                    "type": "string",
                    "example": "all"
                },
                "summary": {
                    "description": "概览统计",
                    "allOf": [
                        {
                            "$ref": "#/definitions/pvesphere_api_v1.DashboardOverviewSummary"
                        }
                    ]
                }
            }
        },
        "pvesphere_api_v1.DashboardOverviewHealth": {
            "type": "object",
            "properties": {
                "critical": {
                    "description": "严重数量",
                    "type": "integer",
                    "example": 0
                },
                "healthy": {
                    "description": "健康数量",
                    "type": "integer",
                    "example": 4
                },
                "warning": {
                    "description": "警告数量",
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "pvesphere_api_v1.DashboardOverviewResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/pvesphere_api_v1.DashboardOverviewData"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "pvesphere_api_v1.DashboardOverviewSummary": {
            "type": "object",
            "properties": {
                "cluster_count": {
                    "description": "集群数量",
                    "type": "integer",
                    "example": 5
                },
                "node_count": {
                    "description": "节点数量",
                    "type": "integer",
                    "example": 42
                },
                "storage_count": {
                    "description": "存储数量",
                    "type": "integer",
                    "example": 18
                },
                "vm_count": {
                    "description": "虚拟机数量",
                    "type": "integer",
                    "example": 1328
                }
            }
        },
        "pvesphere_api_v1.DashboardResourcesData": {
            "type": "object",
            "properties": {
                "cluster_id": {
                    "description": "集群ID（当 scope 为 cluster 时）",
                    "type": "integer"
                },
                "cpu": {
                    "description": "CPU 使用率",
                    "allOf": [
                        {
                            "$ref": "#/definitions/pvesphere_api_v1.ResourceUsage"
                        }
                    ]
                },
                "memory": {
                    "description": "内存使用率",
                    "allOf": [
                        {
                            "$ref": "#/definitions/pvesphere_api_v1.ResourceUsage"
                        }
                    ]
                },
                "scope": {
                    "description": "all 或 cluster",
                    "type": "string",
                    "example": "all"
                },
                "storage": {
                    "description": "存储使用率",
                    "allOf": [
                        {
                            "$ref": "#/definitions/pvesphere_api_v1.ResourceUsage"
                        }
                    ]
                }
            }
        },
        "pvesphere_api_v1.DashboardResourcesResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/pvesphere_api_v1.DashboardResourcesData"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "pvesphere_api_v1.DashboardScopesData": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pvesphere_api_v1.ScopeItem"
                    }
                }
            }
        },
        "pvesphere_api_v1.DashboardScopesResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/pvesphere_api_v1.DashboardScopesData"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "pvesphere_api_v1.GetAccessTicketRequest": {
            "type": "object",
            "required": [
                "cluster_id",
                "password",
                "realm",
                "username"
            ],
            "properties": {
                "cluster_id": {
                    "description": "集群ID（从集群表获取 api_url）",
                    "type": "integer",
                    "example": 1
                },
                "password": {
                    "type": "string",
                    "example": "your-password"
                },
                "realm": {
                    "type": "string",
                    "example": "pam"
                },
                "username": {
                    "type": "string",
                    "example": "root"
                }
            }
        },
        "pvesphere_api_v1.GetAccessTicketResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object",
                    "additionalProperties": true
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "pvesphere_api_v1.NodeHotspots": {
            "type": "object",
            "properties": {
                "cpu": {
                    "description": "CPU Top N",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pvesphere_api_v1.TopResourceConsumer"
                    }
                },
                "memory": {
                    "description": "Memory Top N",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pvesphere_api_v1.TopResourceConsumer"
                    }
                }
            }
        },
...
        "pvesphere_api_v1.OperationItem": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "操作ID",
                    "type": "string",
                    "example": "op-1"
                },
                "name": {
                    "description": "操作名称",
                    "type": "string",
                    "example": "Migrate vm-101 to node-3"
                },
                "operation_type": {
                    "description": "操作类型",
                    "type": "string",
                    "example": "vm_migration"
                },
                "progress": {
                    "description": "进度百分比",
                    "type": "integer",
                    "example": 45
                },
                "started_at": {
                    "description": "开始时间",
                    "type": "string",
                    "example": "2025-12-23T07:50:00Z"
                },
                "status": {
                    "description": "running / pending / failed / completed",
                    "type": "string",
                    "example": "running"
                }
            }
        },
        "pvesphere_api_v1.OperationSummary": {
            "type": "object",
            "properties": {
                "count": {
                    "description": "数量",
                    "type": "integer",
                    "example": 3
                },
                "display_name": {
                    "description": "显示名称",
                    "type": "string",
                    "example": "VM Migrations"
                },
                "operation_type": {
                    "description": "vm_migration / node_maintenance / storage_rebalance",
                    "type": "string",
                    "example": "vm_migration"
                }
            }
        },
        "pvesphere_api_v1.RecentRisk": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "风险ID",
                    "type": "string",
                    "example": "risk-1"
                },
                "level": {
                    "description": "info / warning / critical",
                    "type": "string",
                    "example": "warning"
                },
                "message": {
                    "description": "风险信息",
                    "type": "string",
                    "example": "Node pve-07 unreachable"
                },
                "occurred_at": {
                    "description": "发生时间",
                    "type": "string",
                    "example": "2025-12-23T08:01:00Z"
                },
                "relative_time": {
                    "description": "相对时间",
                    "type": "string",
                    "example": "5 min ago"
                },
                "target_id": {
                    "description": "目标ID",
                    "type": "string",
                    "example": "node-7"
                },
                "target_name": {
                    "description": "目标名称",
                    "type": "string",
                    "example": "pve-07"
                },
                "target_type": {
                    "description": "vm / node / storage / cluster",
                    "type": "string",
                    "example": "node"
                }
            }
        },
...
        "pvesphere_api_v1.ResourceUsage": {
            "type": "object",
            "properties": {
                "total_bytes": {
                    "description": "总字节数（内存/存储）",
                    "type": "integer",
                    "example": 260000000000
                },
                "total_cores": {
                    "description": "CPU 总核心数（仅 CPU）",
                    "type": "number",
                    "example": 250
                },
                "usage_percent": {
                    "description": "使用率百分比",
                    "type": "number",
                    "example": 72
                },
                "used_bytes": {
                    "description": "已使用字节数（内存/存储）",
                    "type": "integer",
                    "example": 180000000000
                },
                "used_cores": {
                    "description": "CPU 已使用核心数（仅 CPU）",
                    "type": "number",
                    "example": 180
                }
            }
        },
...
        "pvesphere_api_v1.ScopeItem": {
            "type": "object",
            "properties": {
                "cluster_id": {
                    "type": "integer",
                    "example": 1
                },
                "cluster_name": {
                    "type": "string",
                    "example": "pve-prod-01"
                },
                "cluster_name_alias": {
                    "type": "string",
                    "example": "生产集群一"
                }
            }
        },
...
        "pvesphere_api_v1.StorageHotspot": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "存储ID",
                    "type": "string",
                    "example": "storage/pve01/local"
                },
                "name": {
                    "description": "存储名称",
                    "type": "string",
                    "example": "local (pve01)"
                },
                "total_bytes": {
                    "description": "总字节",
                    "type": "integer",
                    "example": 100663296000
                },
                "unit": {
                    "description": "单位",
                    "type": "string",
                    "example": "%"
                },
                "usage_percent": {
                    "description": "使用率百分比",
                    "type": "number",
                    "example": 85.5
                },
                "used_bytes": {
                    "description": "已使用字节",
                    "type": "integer",
                    "example": 85899345920
                }
            }
        },
...
        "pvesphere_api_v1.TopResourceConsumer": {
            "type": "object",
            "properties": {
                "cluster_id": {
                    "description": "集群ID",
                    "type": "integer",
                    "example": 1
                },
                "cluster_name": {
                    "description": "集群名称",
                    "type": "string",
                    "example": "cluster1"
                },
                "id": {
                    "description": "资源ID",
                    "type": "string",
                    "example": "qemu/101"
                },
                "metric_value": {
                    "description": "指标值",
                    "type": "number",
                    "example": 92
                },
                "name": {
                    "description": "资源名称",
                    "type": "string",
                    "example": "web-prod-01"
                },
                "node_name": {
                    "description": "节点名称（VM 使用）",
                    "type": "string",
                    "example": "pve01"
                },
                "unit": {
                    "description": "单位",
                    "type": "string",
                    "example": "%"
                }
            }
        },
...
        "pvesphere_api_v1.VMHotspots": {
            "type": "object",
            "properties": {
                "cpu": {
                    "description": "CPU Top N",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pvesphere_api_v1.TopResourceConsumer"
                    }
                },
                "memory": {
                    "description": "Memory Top N",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pvesphere_api_v1.TopResourceConsumer"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}